Changes how PCClassLoader handles files.
Basically, this unwinds a loop that was used for both CLASS: and levels and makes those two separate methods.  This will allow the separation of Class tokens from level tokens in a way that will report errors (6.x behavior) as opposed to unreported problems (5.14 behavior).  This was a necessary prerequisite change to moving the CDOM tokens into PCGen 5.15/5.16


git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk/pcgen@6354 ef53c05b-3124-4d59-ae7d-fe3273f42be7