General tidying up ==================
Tidied up PlayerCharacter getVariable, most of the code was using the same
size constant values, replaced these with a more specific routine.

decrement in VariableProcessor was passed through four or five levels of calls,
but nnever used in any way.  eliminted the variable and the useless parameters. 

Changed the visibility of some operations in PlayerCharacter.


git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk/pcgen@7459 ef53c05b-3124-4d59-ae7d-fe3273f42be7