Reviewed feats Shapeshifter Foil, and the Nightmare tree.
Also working on Monastic Legacy, which changes the monk unarmed strike damage based on non-monk levels. I've implemented a few changes to the _classes and _abilities files to setup the changes required, though these most likely aught to be in the core book. The end result is a new variable defined in the class file called "MonkLevelsAdjusted", which feats or whatnot can add to or subtract from as needed. Also added are a number of new special abilities that are going to be internal in the end, but for now are treated as Special Quality, for testing purposes. These are only added to the monk class if the MonkLevelsAdjusted match. In effect a workaround to setting unarmed strike damage levels. They are in effect no different than the levels added by the core monk class, but letting the UDAM be set not by the core monk level as it is right now. It needs work though. The special abilities gained by having non-monk levels are getting added to the character, but UDAM is not being updated. 
Issue#: NEWSOURCE-130


git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk/pcgen@17608 ef53c05b-3124-4d59-ae7d-fe3273f42be7