Reviewed Haunted Gnome tree*, Hero's Display, Signature Deed, Revelation Strike, and School Strike. Finished reviewing all the feats. I'll need to now go back through and look at the ones that have problems and see if the fixes have already been added or if I can tweak them to work based on any new features or fixes or changes.
Haunted Gnome adds innate spells and increases the number of times some can be cast per day. There's no way as of yet to change this.

Revelation Strike needs work on the chooser. So far it's not showing any chooser or options.

School Strike has a tough chooser and frankly I don't know how to make it work.


Issue#: NEWSOURCE-130


git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk/pcgen@17718 ef53c05b-3124-4d59-ae7d-fe3273f42be7